feat(createStore): export a type for user to contruct action
