+++ test/src/actions.change.ts
@@ -1,13 +1,8 @@
+import { Action } from '../../'
+
 interface State {
   count: number
 }
-
-export type Action<
-  State extends object,
-  Payload = unknown
-> = unknown extends Payload
-  ? <S extends State>(state: S) => S
-  : <S extends State>(state: S, payload: Payload) => S
 
 export const NOCHANGE: Action<State> = (state) => {
   return state
